refactor(web-api): regroup channels by modality and lift non-agent routes
Channels now split by interaction modality (chat, voice) instead of
transport (rest, voice). The chat channel covers turn-based session
delivery — currently WebSocket-transported, but the directory is named
for what it does, not how. Non-agent HTTP routes (admin, billing,
config, content, flashcards, session, wimmelbilder) moved out of
channels/ to web-api/routes/ since they're plain RPC and never talk
to the agent.

Also removed dead shims with no remaining callers: web-api/pipeline/
(stale forks of channels/voice), services/agent_service.py,
services/websocket_service.py, routes/realtime_*.py.
