+++ web-api/channels/chat/__init__.py
@@ -1,0 +1,1 @@
+"""Chat channel — turn-based messages (text, audio, components). Currently WebSocket-transported."""
