--- web-api/channels/rest/__init__.py
@@ -1,1 +1,0 @@
-"""REST and text-WebSocket delivery channel."""
