--- web-api/channels/rest/routes/__init__.py
@@ -1,1 +1,0 @@
-"""REST API route modules."""
