--- web-api/channels/rest/websocket/__init__.py
@@ -1,1 +1,0 @@
-"""Text-mode WebSocket delivery."""
