--- web-api/pipeline/__init__.py
@@ -1,3 +1,0 @@
-"""Backward-compat shim — canonical location is channels.voice.pipeline."""
-
-from channels.voice.pipeline import run_pipecat_agent  # noqa: F401
