--- web-api/routes/realtime_pipecat.py
@@ -1,3 +1,0 @@
-"""Backward-compat shim — canonical location is channels.voice.routes."""
-
-from channels.voice.routes import router  # noqa: F401
