--- web-api/routes/realtime_session.py
@@ -1,3 +1,0 @@
-"""Backward-compat shim — canonical location is channels.rest.websocket.routes."""
-
-from channels.rest.websocket.routes import router  # noqa: F401
