--- web-api/services/agent_service.py
@@ -1,10 +1,0 @@
-"""Backward-compat shim — see harness.runner and harness.turn for the real homes."""
-
-from harness.runner import (  # noqa: F401
-    generate_agent_response,
-    generate_greeting,
-    generate_agent_followup,
-)
-from harness.turn import run_turn  # noqa: F401
-from channels.rest.websocket.session_loop import start_session_loop  # noqa: F401
-from channels.rest.websocket.turn_dispatcher import dispatch_turn as trigger_turn  # noqa: F401
